[REF] hr_recruitment_survey: Put code in proper files
This commit also fixed the bug.

Before - when sharing Recruitment Form from surveys app, additional empty
registration was created.

After this commit, additional empty registration is not created anymore.

task - 2694600
