+++ addons/hr_recruitment_survey/wizard/__init__.py
@@ -1,0 +1,4 @@
+# -*- encoding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import survey_invite
